% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_loc}
\alias{get_loc}
\alias{get_time}
\alias{get_lon}
\alias{get_lat}
\alias{get_lev}
\title{Retrieve geospatial information about a grads resource}
\usage{
get_loc(x, what = "time")

get_time(x, shift = -2)

get_lon(x)

get_lat(x)

get_lev(x)
}
\arguments{
\item{x}{ncdf4 object}

\item{what}{character, one of 'time', 'lon', 'lat', 'lev' or 'all'}

\item{shift}{numeric, see \href{https://github.com/dankelley/oce/issues/738}{oce}}
}
\value{
varies by value of what
\itemize{
  \item{\code{lon, lat, lev} numeric vectors}
  \item{\code{time} POSIXct vector}
  \item{\code{all} list of \code{lon, lat, lev, time} vectors}
}

POSIXct times

numeric longitude locations

numeric latitude locations

numeric level locations
}
\description{
Retrieve geospatial information about a grads resource
}
\section{Functions}{
\itemize{
\item \code{get_time}: retrieve POSIXct times

\item \code{get_lon}: retrieve numeric longitudes

\item \code{get_lat}: retrieve numeric latitudes

\item \code{get_lev}: retrieve numeric levels
}}

