% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_navigation}
\alias{get_navigation}
\title{Retrieve a navigation structure for extracting slabs.}
\usage{
get_navigation(
  x,
  bb = get_bounds(x),
  time = get_loc(x, "time")[1:3],
  lev = get_loc(x, "lev")[1:5]
)
}
\arguments{
\item{x}{ncdf4 object}

\item{bb}{a 4 element bounding box for subsetting ordered as
\code{[xmin, xmax, ymin, ymax]}}

\item{time}{POSIXct vector of one or more times to retrieve. These are matched the
closest known times in the object. See \code{get_time}  Default
is the first recorded time in the object.}

\item{lev}{numeric vector of one or more levels. These are matched the
closest known levels in the object. See \code{get_lev} Default
is the first level time in the object.  Ignored if \code{lev} is not
a dimension of the variable.}
}
\value{
a list of extraction navigation elements
\itemize{
  \item{obb, otime, olev: original request inputs}
  \item{ilon, ilat, itime, ilev: index into dimensions for obb, otime and olev}
  \item{lons, lats, times, levs: full real world dimension values}
  \item{dx, dy: x and y resolution}
  \item{xlim, ylim: limits envelope}
  \item{index: list of array (slab) indices as in [start, count] form}
  \item{stbb: st_bbox object}
  \item{time_index, lev_index: full sequential indices into times and levs}
}
}
\description{
Handy if you are doing multiple extractions - just compute the extraction
navigation once and resuse it (not that computing is slow.)
}
