% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{get_var}
\alias{get_var}
\title{Retrieve a variable as array or stars object.}
\usage{
get_var(
  x,
  var = "tmpsfc",
  bb = get_bounds(x),
  time = get_loc(x, "time")[1:3],
  lev = get_loc(x, "lev")[1:5],
  nav = NULL,
  shift = 2,
  form = c("array", "stars")[2]
)
}
\arguments{
\item{x}{ncdf4 object}

\item{var}{character, one or more names of the variables to retrieve}

\item{bb}{a 4 element bounding box for subsetting ordered as
\code{[xmin, xmax, ymin, ymax]}}

\item{time}{POSIXct vector of one or more times to retrieve. These are matched the
closest known times in the object. See \code{get_time}  Default
is the first recorded time in the object. See \code{shift} argument}

\item{lev}{numeric vector of one or more levels. These are matched the
closest known levels in the object. See \code{get_lev} Default
is the first level time in the object.  Ignored if \code{lev} is not
a dimension of the variable.}

\item{nav}{list, see \code{get_navigation} If not provided then computed from
\code{bb, time and lev}.  If provided then \code{bb, time and lev} are
ignored (since you have already computed the navigation.)}

\item{shift}{2 number of days to shift the input time - see \code{get_time}. Set
to \code{0} to add no shift}

\item{form}{character either 'array' of 'stars' (default)
\itemize{
  \item{array}{an array or list of arrays, possibly degenerate to a matrix}
  \item{stars}{a stars object, possibly with band (time) and z (level)}
}}
}
\description{
Data are stored as \code{[lon, lat, time]} or \code{[lon, lat, lev, time]}
Degenerate indices (dimension = 1) are discarded, so if a single time is
requested for a \code{[lon, lat, time]} variable then a single band object is
returned.
}
\details{
The requested bounding box coordinates are matched to the closest grid cell
centers, thus the output grids may differ in extent form the requested bounding
box.

Requested times and levels are considered contiguous - we are extracting slabs
of data after all. Currently the first and last times or levels requested mark
the inclusive bounds of the slab in those dimensions. Requesting a single time or
level works pefectly well.  If you need disjoint bands (not contiguous bands) then
you will need to make a separate request for each.
}
