% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomads-ncep.R
\name{grads_base_uri}
\alias{grads_base_uri}
\title{Retrieve the base uri for either opendap or html service}
\usage{
grads_base_uri(what = c("opendap", "html")[1])
}
\arguments{
\item{what}{character, the name of the desired service}
}
\value{
URI
}
\description{
Retrieve the base uri for either opendap or html service
}
