% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomads-ncep.R
\name{grads_uri}
\alias{grads_uri}
\title{Generate an opendap uri}
\usage{
grads_uri(
  date = Sys.Date(),
  product = "nam_00z",
  base_uri = grads_base_uri(what = "opendap")
)
}
\arguments{
\item{date}{Date or character, the date to query}

\item{product}{character, the product to query}

\item{base_uri, }{character, the base grads uri}
}
\description{
Generate an opendap uri
}
