% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nc.R
\name{open_grads}
\alias{open_grads}
\title{Open an opendap resource}
\usage{
open_grads(uri, tries = 3, sleep = 3)
}
\arguments{
\item{uri}{character, the uri for the resource}

\item{tries}{numeric, number of tries before failing}

\item{sleep}{numeric, seconds to pause between attempts to connect}
}
\value{
ncdf4 object or NULL
}
\description{
Open an opendap resource
}
